Revised solution for day04
Updated how I parsed the winning and rolled numbers to not require as
much splitting and conditions to check for empty strings. Also added a
helper function to compare results with intersecting sets. Also Updated
solution 2 to use an iterative solution that is a bit faster than the
recusive solution I had before.